Add Database.PostgreSQL.Tagged.
This tagged-effect module allows you to represent the semantics of your
queries (i.e. whether they are idempotent or not) in the Haskell type
system. This can be useful when dealing with Postgres replicas:
read-only queries can be farmed out to replicas, whereas queries that
write to the database can be dispatched directly to the primary.