Add a MonadCatch instance with transaction savepoints.
This implements catching behavior that is transactional utilizing savepoints. The implementation is not trivial to ensure prompt releasing of savepoints, so in addition to this instance a test suite has been added. It adds a few sanity tests for the basic behavior of PGTransaction and tests specific to MonadCatch instance. (#21)

This implements issue #20